libgo: merge gc go1.25.0 -> go1.26.0 (stdlib+runtime), resolve conflicts
Merge via /tmp/merge126.sh (OLDDIR=/tmp/go125 go1.25.0, NEWDIR=/tmp/go126
go1.26.0=d90b98e653; VERSION=go1.26.0, MERGE=d90b98e653...). 86 conflicts:
9 build-tag-only auto-resolved, 77 substantive resolved by 6 parallel agents
by area (guide /tmp/gccgo_merge_guide_126.md). Philosophy: keep gccgo TOP for
runtime/reflect/syscall fork (libffi, libc/pthreads, own hmap/bmap maps, C
atomics, no gc asm/pclntab/abi.RegArgs); take gc for portable library/cmd.
gc 1.26 changes REJECTED: new _mkmalloc malloc codegen, green-tea GC,
goroutine-leak detection, swiss maps, heap-base randomization, arenas,
register ABI. This is the merge checkpoint; runtime build reconciliation
(e.g. abi.OldMapBucketCount rename) follows in subsequent commits.

+++ libgo/go/internal/runtime/sys/intrinsics.go
@@ -140,7 +140,7 @@ func Len64(x uint64) (n int) {
 		x >>= 8
 		n += 8
 	}
-	return n + int(len8tab[x])
+	return n + int(len8tab[uint8(x)])
 }
 
 // --- OnesCount ---
